requirements.txt
nextcord>=2.4.1
python-decouple>=3.7